requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
